chore: revert  "chore: Optimize hardware interaction prompts (#986)"
This reverts commit efac73b62f26a9590725ce347d919adaa8f42c50.
