+++ packages/kit/src/utils/hardware/ui-response.tsx
@@ -3,15 +3,13 @@ import React from 'react';
 import { UI_RESPONSE } from '@onekeyfe/hd-core';
 import { PermissionsAndroid } from 'react-native';
 
-import { DialogManager } from '@onekeyhq/components';
+import { DialogManager, ToastManager } from '@onekeyhq/components';
 import PermissionDialog from '@onekeyhq/kit/src/components/PermissionDialog/PermissionDialog';
 import { navigationRef } from '@onekeyhq/kit/src/provider/NavigationProvider';
 import platformEnv from '@onekeyhq/shared/src/platformEnv';
 
 import deviceUtils from './deviceUtils';
 import { getHardwareSDKInstance } from './hardwareInstance';
-import RequestConfirmView from './RequestView/RequestConfirm';
-import RequestPinView from './RequestView/RequestPin';
 
 export const UI_REQUEST = {
   REQUEST_PIN: 'ui-request_pin',
@@ -33,28 +31,15 @@ export const UIResponse = async (event: any, hideUiResponse?: boolean) => {
 
   switch (type) {
     case UI_REQUEST.REQUEST_PIN:
-      DialogManager.hide();
-      DialogManager.show({
-        render: (
-          <RequestPinView
-            deviceType={deviceUtils?.connectedDeviceType}
-            onCancel={async () => {
-              // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-              const device = event.payload?.device || {};
-              const { connectId } = device || {};
-
-              HardwareSDK.cancel(connectId);
-              await deviceUtils.getFeatures(connectId);
-            }}
-            onConfirm={(pin: string) => {
-              HardwareSDK.uiResponse({
-                type: UI_RESPONSE.RECEIVE_PIN,
-                payload: pin,
-              });
-            }}
-          />
-        ),
-      });
+      ToastManager.show(
+        {
+          deviceType: deviceUtils?.connectedDeviceType,
+        },
+        {
+          autoHide: false,
+          type: 'enterPinOnDevice',
+        },
+      );
       if (hideUiResponse) {
         return;
       }
@@ -65,32 +50,25 @@ export const UIResponse = async (event: any, hideUiResponse?: boolean) => {
       break;
 
     case UI_REQUEST.INVALID_PIN: {
-      DialogManager.hide();
+      ToastManager.hide();
       navigationRef.current?.goBack?.();
       break;
     }
 
     case UI_REQUEST.REQUEST_BUTTON:
-      DialogManager.hide();
-      DialogManager.show({
-        render: (
-          <RequestConfirmView
-            deviceType={deviceUtils?.connectedDeviceType}
-            onCancel={async () => {
-              // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-              const device = event.payload?.device || {};
-              const { connectId } = device || {};
-
-              HardwareSDK.cancel(connectId);
-              await deviceUtils.getFeatures(connectId);
-            }}
-          />
-        ),
-      });
+      ToastManager.show(
+        {
+          deviceType: deviceUtils?.connectedDeviceType,
+        },
+        {
+          autoHide: false,
+          type: 'confirmOnDevice',
+        },
+      );
       break;
 
     case UI_REQUEST.CLOSE_UI_WINDOW: {
-      DialogManager.hide();
+      ToastManager.hide();
       break;
     }
     case UI_REQUEST.LOCATION_PERMISSION:
